+++ src/main/groovy/nextflow/validation/samplesheet/SamplesheetConverter.groovy
@@ -101,7 +101,7 @@ class SamplesheetConverter {
         // Validate
         final validator = new JsonSchemaValidator(config)
         def JSONArray samplesheet = fileToJson(samplesheetFile, schemaFile) as JSONArray
-        def Tuple2<List<String>,List<String>> validationResults = validator.validate(samplesheet, schemaFile.text)
+        def Tuple2<List<String>,List<String>> validationResults = validator.validate(samplesheet, schemaFile.text, schemaFile.toString())
         def validationErrors = validationResults[0]
         if (validationErrors) {
             def msg = "${colors.red}The following errors have been detected in ${samplesheetFile.toString()}:\n\n" + validationErrors.join('\n').trim() + "\n${colors.reset}\n"